+++ trinity/common/config_validator.py
@@ -21,7 +21,7 @@ from trinity.utils.log import get_logger
 from trinity.utils.lora_utils import create_dummy_lora
 
 if TYPE_CHECKING:
-    from trinity.trainer.verl.verl_config import FSDPConfig
+    from trinity.trainer.verl_legacy.verl_config import FSDPConfig
 
 
 class ConfigValidator(ABC):
@@ -1227,51 +1227,50 @@ class TrainerConfigValidator(ConfigValidator):
         config.trainer.trust_remote_code = config.model.trust_remote_code
 
         if config.trainer.trainer_type == "verl":
-            if config.trainer.ulysses_sequence_parallel_size < 1:
-                self.logger.warning(
-                    "Ulysses sequence parallel size is set to 1 "
-                    f"because {config.trainer.ulysses_sequence_parallel_size} is invalid."
-                )
-                config.trainer.ulysses_sequence_parallel_size = 1
+            from trinity.trainer.trainer import is_verl_legacy
 
-            if config.trainer.trainer_config:
-                from trinity.trainer.verl.verl_config import veRLConfig
+            if is_verl_legacy():
+                if config.trainer.ulysses_sequence_parallel_size < 1:
+                    self.logger.warning(
+                        "Ulysses sequence parallel size is set to 1 "
+                        f"because {config.trainer.ulysses_sequence_parallel_size} is invalid."
+                    )
+                    config.trainer.ulysses_sequence_parallel_size = 1
 
-                trainer_config_schema = OmegaConf.structured(veRLConfig)
-                trainer_config = OmegaConf.merge(
-                    trainer_config_schema, config.trainer.trainer_config
-                )
-                config.trainer.trainer_config = OmegaConf.to_object(trainer_config)
-            elif config.trainer.trainer_config_path:
-                raise ValueError(
-                    "`trainer_config_path` is deprecated; please use `trainer_config` instead."
-                )
+                if config.trainer.max_token_len_per_gpu is None:
+                    config.trainer.max_token_len_per_gpu = math.ceil(
+                        2 * config.model.max_model_len / config.trainer.ulysses_sequence_parallel_size  # type: ignore [operator]
+                    )
+
+                if config.trainer.trainer_config:
+                    from trinity.trainer.verl_legacy.verl_config import veRLConfig
+
+                    trainer_config_schema = OmegaConf.structured(veRLConfig)
+                    trainer_config = OmegaConf.merge(
+                        trainer_config_schema, config.trainer.trainer_config
+                    )
+                    config.trainer.trainer_config = OmegaConf.to_object(trainer_config)
+                elif config.trainer.trainer_config_path:
+                    raise ValueError(
+                        "`trainer_config_path` is deprecated; please use `trainer_config` instead."
+                    )
+                else:
+                    from trinity.trainer.verl_legacy.verl_config import veRLConfig
+
+                    self.logger.info(
+                        "`trainer_config` is not provided, using default trainer config."
+                    )
+                    config.trainer.trainer_config = veRLConfig()
+                config.trainer.trainer_config.synchronize_config(config)
             else:
-                from trinity.trainer.verl.verl_config import veRLConfig
+                config.trainer.trainer_config = None
+                if config.trainer.ulysses_sequence_parallel_size < 1:
+                    config.trainer.ulysses_sequence_parallel_size = 1
 
-                self.logger.info("`trainer_config` is not provided, using default trainer config.")
-                config.trainer.trainer_config = veRLConfig()
-            if config.trainer.max_token_len_per_gpu is None:
-                config.trainer.max_token_len_per_gpu = math.ceil(
-                    2 * config.model.max_model_len / config.trainer.ulysses_sequence_parallel_size  # type: ignore [operator]
-                )
-            if config.trainer.save_hf_checkpoint not in {"last", "always", "never"}:
-                raise ValueError(
-                    f"Invalid trainer.save_hf_checkpoint: {config.trainer.save_hf_checkpoint}, "
-                    "must be one of 'last', 'always', or 'never'."
-                )
-            config.trainer.trainer_config.synchronize_config(config)
-        elif config.trainer.trainer_type == "verl08":
-            # verl08 uses build_verl_config() at trainer init time, so no
-            # trainer_config preprocessing is needed here.  All fields are
-            # read directly from the Trinity Config dataclass.
-            config.trainer.trainer_config = None
-            if config.trainer.ulysses_sequence_parallel_size < 1:
-                config.trainer.ulysses_sequence_parallel_size = 1
-            if config.trainer.max_token_len_per_gpu is None:
-                config.trainer.max_token_len_per_gpu = math.ceil(
-                    2 * config.model.max_model_len / config.trainer.ulysses_sequence_parallel_size  # type: ignore [operator]
-                )
+                if config.trainer.max_token_len_per_gpu is None:
+                    config.trainer.max_token_len_per_gpu = math.ceil(
+                        2 * config.model.max_model_len / config.trainer.ulysses_sequence_parallel_size  # type: ignore [operator]
+                    )
             if config.trainer.save_hf_checkpoint not in {"last", "always", "never"}:
                 raise ValueError(
                     f"Invalid trainer.save_hf_checkpoint: {config.trainer.save_hf_checkpoint}, "
@@ -1481,7 +1480,7 @@ class GPUMemoryValidator(ConfigValidator):
         Raises:
             ValueError: If estimated memory usage exceeds safe limits and suggestions are not bypassed.
         """
-        from trinity.trainer.verl.verl_config import veRLConfig
+        from trinity.trainer.verl_legacy.verl_config import veRLConfig
 
         self.pytorch_env_flag = (
             os.environ.get("PYTORCH_CUDA_ALLOC_CONF", "") == "expandable_segments:True"
