+++ trinity/trainer/__init__.py
@@ -1,3 +1,8 @@
-from trinity.trainer.trainer import TrainEngineWrapper, Trainer, get_trainer_wrapper
+from trinity.trainer.trainer import (
+    TrainEngineWrapper,
+    Trainer,
+    get_trainer_wrapper,
+    is_verl_legacy,
+)
 
-__all__ = ["Trainer", "TrainEngineWrapper", "get_trainer_wrapper"]
+__all__ = ["Trainer", "TrainEngineWrapper", "get_trainer_wrapper", "is_verl_legacy"]
