+++ trinity/trainer/verl/__init__.py
@@ -1,0 +1,1 @@
+"""Verl v0.8.0 trainer"""
