--- trinity/trainer/verl08/__init__.py
@@ -1,1 +1,0 @@
-"""Verl v0.8.0 trainer"""
